step all done, plus db fixed, admin added user-strong_password

+++ ads/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Ads(models.Model):
-    id = models.PositiveBigIntegerField(primary_key=True)
+    id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=140)
     author = models.CharField(max_length=140)
     price = models.DecimalField(max_digits=7, decimal_places=2)
@@ -12,5 +12,5 @@ class Ads(models.Model):
 
 
 class Categories(models.Model):
-    id = models.PositiveBigIntegerField(primary_key=True)
+    id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=140)
